pyproject.toml
[project]
name = "vlm-anchor-project"
version = "0.1.0"
description = "Cross-modal anchoring experiment code for VQA number questions with irrelevant images."
requires-python = ">=3.10"
dependencies = [
  "accelerate>=1.10.0",
  "diffusers>=0.36.0",
  "einops>=0.8.0",
  "huggingface_hub>=0.34.0",
  "matplotlib>=3.9.0",
  "numpy>=1.26.0",
  "ollama>=0.5.1",
  "pandas>=2.2.0",
  "pillow>=10.4.0",
  "protobuf>=5.29.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.0",
  "safetensors>=0.6.0",
  "scikit-image>=0.24.0",
  "sentencepiece>=0.2.0",
  "tiktoken>=0.9.0",
  "timm>=1.0.15",
  "torch>=2.3.0",
  "torchvision>=0.18.0",
  "transformers>=4.57.0",
  "qwen-vl-utils>=0.0.14",
  "tqdm>=4.66.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
